Extracted url conversion to separate class
