+++ src/phpDocumentor/Transformer/Router/UrlGenerator/Standard/FileDescriptor.php
@@ -25,36 +25,8 @@ class FileDescriptor implements UrlGeneratorInterface
      */
     public function __invoke($node)
     {
-        return '/files/' . $this->convertFilePathToUrl($node->getPath()) .'.html';
-    }
+        $converter = new QualifiedNameToUrlConverter();
 
-    /**
-     * Converts the given path to a valid url.
-     *
-     * @param string $path
-     *
-     * @return string
-     */
-    private function convertFilePathToUrl($path)
-    {
-        $path = $this->removeFileExtensionFromPath($path);
-
-        return str_replace(array('/', '\\'), '.', ltrim($path, '/'));
-    }
-
-    /**
-     * Removes the file extension from the provided path.
-     *
-     * @param string $path
-     *
-     * @return string
-     */
-    private function removeFileExtensionFromPath($path)
-    {
-        if (strrpos($path, '.') !== false) {
-            $path = substr($path, 0, strrpos($path, '.'));
-        }
-
-        return $path;
+        return '/files/' . $converter->fromFile($node->getPath()) .'.html';
     }
 }
