+++ src/phpDocumentor/Transformer/Router/UrlGenerator/Standard/FunctionDescriptor.php
@@ -28,27 +28,8 @@ class FunctionDescriptor implements UrlGeneratorInterface
      */
     public function __invoke($node)
     {
-        $namespaceName = $node->getNamespace();
+        $converter = new QualifiedNameToUrlConverter();
 
-        return '/namespaces/' . $this->convertFqcnToFilename($namespaceName) . '.html#function_' . $node->getName();
-    }
-
-    /**
-     * Converts the provided FQCN into a file name by replacing all slashes with dots.
-     *
-     * @param string $fqcn
-     *
-     * @return string
-     */
-    private function convertFqcnToFilename($fqcn)
-    {
-        $name = str_replace('\\', '.', ltrim($fqcn, '\\'));
-
-        // convert root namespace to default; default is a keyword and no namespace CAN be named as such
-        if ($name === '') {
-            $name = 'default';
-        }
-
-        return $name;
+        return '/namespaces/' . $converter->fromNamespace($node->getNamespace()) . '.html#function_' . $node->getName();
     }
 }
