+++ src/phpDocumentor/Transformer/Router/UrlGenerator/Standard/PackageDescriptor.php
@@ -25,26 +25,8 @@ class PackageDescriptor implements UrlGeneratorInterface
      */
     public function __invoke($node)
     {
-        return '/packages/' . $this->convertFqcnToFilename($node->getFullyQualifiedStructuralElementName()) .'.html';
+        $converter = new QualifiedNameToUrlConverter();
+
+        return '/packages/' . $converter->fromPackage($node->getFullyQualifiedStructuralElementName()) .'.html';
     }
-
-    /**
-     * Converts the provided FQCN into a file name by replacing all slashes with dots.
-     *
-     * @param string $fqcn
-     *
-     * @return string
-     */
-    private function convertFqcnToFilename($fqcn)
-    {
-        $name = str_replace(array('\\', '_'), '.', ltrim($fqcn, '\\'));
-
-        // convert root namespace to default; default is a keyword and no namespace CAN be named as such
-        if ($name === '') {
-            $name = 'default';
-        }
-
-        return $name;
-    }
-
 }
