+++ tests/unit/phpDocumentor/Transformer/Router/UrlGenerator/Standard/PropertyDescriptorTest.php
@@ -20,7 +20,7 @@ class PropertyDescriptorTest extends \PHPUnit_Framework_TestCase
 {
     /**
      * @covers phpDocumentor\Transformer\Router\UrlGenerator\Standard\PropertyDescriptor::__invoke
-     * @covers phpDocumentor\Transformer\Router\UrlGenerator\Standard\PropertyDescriptor::convertFqcnToFilename
+     * @covers phpDocumentor\Transformer\Router\UrlGenerator\Standard\QualifiedNameToUrlConverter::fromClass
      */
     public function testGenerateUrlForPropertyDescriptor()
     {
